@@ -1,20 +1,19 @@
 import { TestBed } from '@angular/core/testing';
 
 import {ApiClientService} from "./apiclient.service";
-import {HttpClient, HttpClientModule} from "@angular/common/http";
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 
 describe('ApiclientService', () => {
   let service: ApiClientService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [
-        HttpClientModule
-      ],
-      providers:[
-        HttpClient
-      ]
-    });
+    imports: [],
+    providers: [
+        HttpClient,
+        provideHttpClient(withInterceptorsFromDi())
+    ]
+});
     service = TestBed.inject(ApiClientService);
   });
 
