@@ -1,5 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-import {HttpClientModule} from "@angular/common/http";
+import { provideHttpClient, withInterceptorsFromDi } from "@angular/common/http";
 import { TodoService } from './todo.service';
 
 describe('TodoService', () => {
@@ -7,8 +7,9 @@ describe('TodoService', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientModule],
-    });
+    imports: [],
+    providers: [provideHttpClient(withInterceptorsFromDi())]
+});
     service = TestBed.inject(TodoService);
   });
 
